// src/enums/FileUpload.ts
export interface FileData {
  filePath: string,

  mimeType: string;
  sizeInBytes: number;
}


export enum FileCloudProviders {
  CLOUDINARY = 'CLOUDINARY',
}


export enum MortgageAccountVerificationFiles {
  BANK_STATEMENT = "BANK_STATEMENT",
  GOVERNMENT_APPROVED_ID = "GOVERNMENT_APPROVED_ID",
  RECENT_UTILITY_BILL = "RECENT_UTILITY_BILL",
  CAC_CERTIFICATE = "CAC_CERTIFICATE"  
} 

export enum UploadFileCategory {
  CATEGORY_PHOTO = 'CATEGORY_PHOTO',
  CATEGORY_BANNER = 'CATEGORY_BANNER',
  PRODUCT_PHOTO = 'PRODUCT_PHOTO',
  BRAND_PHOTO = 'BRAND_PHOTO',
  USER_PHOTO = 'USER_PHOTO',
  STORE_FRONT_BANNER = 'STORE_FRONT_BANNER',
  PROCURMENT_LIST = 'PROCURMENT_LIST',

  PRODUCT_LEASE_REQUEST_ID_CARD = 'PRODUCT_LEASE_REQUEST_ID_CARD',
  PRODUCT_LEASE_REQUEST_CAC_CERTIFICATE = 'PRODUCT_LEASE_REQUEST_CAC_CERTIFICATE',
  PRODUCT_LEASE_BANK_STATEMENT = 'PRODUCT_LEASE_BANK_STATEMENT',
  PRODUCT_LEASE_UTILITY_BILL = 'PRODUCT_LEASE_UTILITY_BILL',
  PRODUCT_LEASE_DISTRIBUTORSHIP_APPOINTMENT_LETTER = 'PRODUCT_LEASE_DISTRIBUTORSHIP_APPOINTMENT_LETTER',

  SELLER_CAC_DOCUMENT = 'SELLER_CAC_DOCUMENT',
  SELLER_ID_CARD = 'SELLER_ID_CARD',
  SELLER_COMPANY_LOGO = 'SELLER_COMPANY_LOGO',

  BULK_PRODUCTS_FILE = 'BULK_PRODUCTS_FILE',
  USER_BACK_DROP_PHOTO = "USER_BACK_DROP_PHOTO",
  PROJECT_IMAGES = "PROJECT_IMAGES",
  BANK_STATEMENT = "BANK_STATEMENT",
  GOVERNMENT_APPROVED_ID = "GOVERNMENT_APPROVED_ID",
  RECENT_UTILITY_BILL = "RECENT_UTILITY_BILL",
  CAC_CERTIFICATE = "CAC_CERTIFICATE" 
}